requirements.txt
Flask
authlib
python-dotenv
earthengine-api